refactor(js/plugins/vertexai): extract common logic and types
