+++ js/plugins/vertexai/src/gemini.ts
@@ -46,7 +46,7 @@ import {
   downloadRequestMedia,
   simulateSystemPrompt,
 } from 'genkit/model/middleware';
-import { PluginOptions } from './index.js';
+import { PluginOptions } from './common/types';
 
 const SafetySettingsSchema = z.object({
   category: z.nativeEnum(HarmCategory),
